Refactor billing and compliance utilities; consolidate common functions
- Removed redundant utility functions from various services and replaced them with centralized implementations in `compliance-utils.ts` and `display-utils.ts`.
- Updated `applyGuestRetentionPolicy` and `applyBillingRetentionPolicy` to use the new utility functions for redacting sensitive information.
- Simplified the cashier session handling by delegating to a shared implementation in `command-consumer-utils`.
- Enhanced tenant scope handling in checkout routes to ensure proper RLS configuration.
- Improved room query handling by introducing a `RoomQueryRow` type and a mapping utility for better code reuse.
- Added new utility functions for formatting enum values and coercing timestamps to ISO strings.
- Updated SQL scripts to ensure proper handling of tenant isolation policies.
