+++ Apps/core-service/src/routes/settings-packages.ts
@@ -22,6 +22,7 @@ import {
   updatePackage,
 } from "../repositories/packages-repository.js";
 import type { AuthUser } from "../types/auth.js";
+import { hasScope } from "../utils/scope.js";
 
 const PACKAGES_TAG = "Packages";
 
@@ -111,17 +112,6 @@ const createComponentBody: JsonSchema = schemaFromZod(
 // =====================================================
 // AUTH HELPERS
 // =====================================================
-
-const hasScope = (user: AuthUser | undefined, requiredScope: string) => {
-  if (!user) {
-    return false;
-  }
-  if (!user.scope) {
-    return false;
-  }
-  const normalizedScopes = Array.isArray(user.scope) ? user.scope : user.scope.split(" ");
-  return normalizedScopes.includes(requiredScope);
-};
 
 const enforceScope = (
   request: FastifyRequest,
