+++ Apps/rooms-service/src/sources/available-rooms-source.ts
@@ -9,6 +9,7 @@ import { BaseSource, type PipelineContext } from "@tartware/candidate-pipeline";
 
 import { query } from "../lib/db.js";
 import type { RoomCandidate, RoomRecommendationQuery } from "../types.js";
+import { type RoomQueryRow, rowToRoomCandidate } from "./room-query-utils.js";
 
 export class AvailableRoomsSource extends BaseSource<RoomRecommendationQuery, RoomCandidate> {
   readonly name = "available_rooms";
@@ -31,17 +32,7 @@ export class AvailableRoomsSource extends BaseSource<RoomRecommendationQuery, Ro
     // 2. Have status 'available' or 'clean'
     // 3. Are not locked/reserved for the requested dates
     // 4. Can accommodate the guest count
-    const result = await query<{
-      room_id: string;
-      room_type_id: string;
-      room_type_name: string;
-      room_number: string;
-      floor: string;
-      view_type: string | null;
-      base_rate: number;
-      status: string;
-      max_occupancy: number;
-    }>(
+    const result = await query<RoomQueryRow>(
       `
       SELECT
         r.id AS room_id,
@@ -82,18 +73,9 @@ export class AvailableRoomsSource extends BaseSource<RoomRecommendationQuery, Ro
       ],
     );
 
-    const candidates: RoomCandidate[] = result.rows.map((row) => ({
-      roomId: row.room_id,
-      roomTypeId: row.room_type_id,
-      roomTypeName: row.room_type_name,
-      roomNumber: row.room_number,
-      floor: row.floor,
-      viewType: row.view_type ?? undefined,
-      baseRate: Number(row.base_rate),
-      status: row.status,
-      maxOccupancy: row.max_occupancy,
-      source: this.name,
-    }));
+    const candidates: RoomCandidate[] = result.rows.map((row) =>
+      rowToRoomCandidate(row, this.name),
+    );
 
     context.logger.debug({ count: candidates.length }, "Available rooms fetched");
     return candidates;
